+++ src/ir_RC5_RC6.hpp
@@ -66,7 +66,7 @@ uint8_t sLastSendToggleValue = 1; // To start first command with toggle 0
 
 #define RC5_DURATION        (15L * RC5_UNIT) // 13335
 #define RC5_REPEAT_PERIOD   (128L * RC5_UNIT) // 113792
-#define RC5_REPEAT_SPACE    (RC5_REPEAT_PERIOD - RC5_DURATION) // 100 ms
+#define RC5_REPEAT_DISTANCE (RC5_REPEAT_PERIOD - RC5_DURATION) // 100 ms
 
 /************************************
  * Start of send and decode functions
@@ -110,7 +110,7 @@ void IRsend::sendRC5(uint8_t aAddress, uint8_t aCommand, int_fast8_t aNumberOfRe
         // skip last delay!
         if (tNumberOfCommands > 0) {
             // send repeated command in a fixed raster
-            delay(RC5_REPEAT_SPACE / MICROS_IN_ONE_MILLI);
+            delay(RC5_REPEAT_DISTANCE / MICROS_IN_ONE_MILLI);
         }
     }
     IrReceiver.restartAfterSend();
@@ -194,7 +194,7 @@ bool IRrecv::decodeRC5() {
     decodedIRData.protocol = RC5;
 
     // check for repeat
-    checkForRepeatSpaceAndSetFlag(RC5_REPEAT_SPACE / MICROS_IN_ONE_MILLI);
+    checkForRepeatSpaceAndSetFlag(RC5_REPEAT_DISTANCE / MICROS_IN_ONE_MILLI);
 
     return true;
 }
@@ -233,7 +233,7 @@ bool IRrecv::decodeRC5() {
 #define RC6_TRAILING_SPACE  (6 * RC6_UNIT) // 2666
 #define MIN_RC6_MARKS       4 + ((RC6_ADDRESS_BITS + RC6_COMMAND_BITS) / 2) // 12, 4 are for preamble
 
-#define RC6_REPEAT_SPACE    107000 // just a guess but > 2.666ms
+#define RC6_REPEAT_DISTANCE 107000 // just a guess but > 2.666ms
 
 /**
  * Main RC6 send function
@@ -337,7 +337,7 @@ void IRsend::sendRC6(uint8_t aAddress, uint8_t aCommand, int_fast8_t aNumberOfRe
         // skip last delay!
         if (tNumberOfCommands > 0) {
             // send repeated command in a fixed raster
-            delay(RC6_REPEAT_SPACE / MICROS_IN_ONE_MILLI);
+            delay(RC6_REPEAT_DISTANCE / MICROS_IN_ONE_MILLI);
         }
     }
 }
@@ -451,7 +451,7 @@ bool IRrecv::decodeRC6() {
     }
 
     // check for repeat, do not check toggle bit yet
-    if (decodedIRData.rawDataPtr->rawbuf[0] < ((RC6_REPEAT_SPACE + (RC6_REPEAT_SPACE / 4)) / MICROS_PER_TICK)) {
+    if (decodedIRData.rawDataPtr->rawbuf[0] < ((RC6_REPEAT_DISTANCE + (RC6_REPEAT_DISTANCE / 4)) / MICROS_PER_TICK)) {
         decodedIRData.flags |= IRDATA_FLAGS_IS_REPEAT;
     }